small adjustments on the ui to update DB secrets list

+++ backend/src/services/pipeline/DltPipeline.py
@@ -281,7 +281,8 @@ class DltPipeline:
         
 
     def revert_ppline(self):
-        os.remove(self.curr_file)
+        if(type(self.curr_file) == str):
+            os.remove(self.curr_file)
 
 
     @staticmethod
